backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.0.4
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.52.2
httpx==0.27.2
websockets==13.1
aiosmtplib==3.0.1
emails==0.6
